feat(backend): add auth middleware to ticketRoutes

+++ server/src/router/ticketRoutes.ts
@@ -1,6 +1,10 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import { CustomRequest } from '../types/CustomRequest';
 import prisma from '../lib/prisma';
+import { GlobalRole, ProjectRole } from '@prisma/client';
+import { authorizeGlobalRole } from '../middleware/authorizeGlobalRole';
+import { checkProjectMembership } from '../middleware/checkProjectMembership';
+import { checkProjectRole } from '../middleware/checkProjectRole';
 import {
   getAllTickets,
   getTicketById,
@@ -10,17 +14,25 @@ import {
   deleteTicket,
   getTicketsByBoardId,
 } from '../controllers/ticketController';
+import { checkTicketOwnership } from '../middleware/checkTicketOwnership';
 
 const router = Router();
 
 // Get all Tickets
-router.get('/tickets', async (req: Request, res: Response): Promise<void> => {
-  await getAllTickets(req, res, prisma);
-});
+router.get(
+  '/tickets',
+  authorizeGlobalRole(GlobalRole.ADMIN),
+  async (req: Request, res: Response): Promise<void> => {
+    await getAllTickets(req, res, prisma);
+  }
+);
 
 // Get Ticket by Id
 router.get(
   '/tickets/:id',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.ADMIN),
   async (req: Request, res: Response): Promise<void> => {
     await getTicketById(req, res, prisma);
   }
@@ -29,6 +41,9 @@ router.get(
 // Get all Tickets by User Id
 router.get(
   '/tickets/assigneeId/:userId',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   async (req: Request, res: Response): Promise<void> => {
     await getTicketByAssigneeId(req, res, prisma);
   }
@@ -37,6 +52,9 @@ router.get(
 // Get Tickets by Board Id
 router.get(
   '/tickets/:boardId/board',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.VIEWER),
   async (req: Request, res: Response): Promise<void> => {
     await getTicketsByBoardId(req, res, prisma);
   }
@@ -45,6 +63,9 @@ router.get(
 // Create new ticket
 router.post(
   '/tickets',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await createNewTicket(req as CustomRequest, res, next, prisma);
   }
@@ -53,6 +74,10 @@ router.post(
 // Delete ticket
 router.delete(
   '/tickets/:id',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
+  checkTicketOwnership(prisma),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await deleteTicket(req as CustomRequest, res, next, prisma);
   }
@@ -60,7 +85,11 @@ router.delete(
 
 // Update a ticket
 router.patch(
-  '/tickets/updateTicket/:ticketId',
+  '/tickets/updateTicket/:id',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
+  checkTicketOwnership(prisma),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await updateTicket(req as CustomRequest, res, next, prisma);
   }
